backend/requirements.txt
grpcio==1.57.0
grpcio-tools==1.57.0
psycopg2-binary==2.9.9
numpy~=2.0.2
joblib~=1.4.2
protobuf~=4.25.5
kafka-python>=2.0.2
pandas~=2.2.3
scikit-learn~=1.5.2
redis~=5.2.1
bcrypt~=4.2.1
jwt~=1.3.1
python-dotenv~=1.0.1